Added policies and abilities for task management

+++ app/Http/Controllers/Api/V1/TaskController.php
@@ -10,8 +10,10 @@ use App\Http\Requests\Api\V1\UpdateTaskRequest;
 use App\Http\Resources\V1\TaskResource;
 use App\Models\Task;
 use App\Models\User;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\Request;
+use PhpParser\Builder\Class_;
 
 class TaskController extends ApiController
 {
@@ -29,12 +31,14 @@ class TaskController extends ApiController
     public function store(StoreTaskRequest $request)
     {
         try {
-            $user = User::findOrFail($request->input('data.relationships.author.data.id'));
-        } catch (ModelNotFoundException $exception) {
-            return $this->error('User not found', 404);
-        }
+            
+            $this->authorize('store', Task::class);
 
-        return new TaskResource(Task::create($request->mappedAttributes()));
+            return new TaskResource(Task::create($request->mappedAttributes()));
+        
+        } catch (AuthorizationException $exception) {
+            return $this->error('You are not authorized to create this task.', 403);
+        } 
     }
 
     /**
@@ -50,7 +54,7 @@ class TaskController extends ApiController
             }
             
             return new TaskResource($task);
-        }  catch (ModelNotFoundException $exception) {
+        }   catch (ModelNotFoundException $exception) {
             return $this->error('Task cannot be found.', 404);
         }
     }
@@ -62,13 +66,18 @@ class TaskController extends ApiController
     {
         try {
             $task = Task::findOrFail($task_id);
+
+            $this->authorize('update', $task);
             
             $task->update($request->mappedAttributes());
     
             return new TaskResource($task);
         } catch (ModelNotFoundException $exception) {
-            return $this->error('Ticket cannot be found.', 404);
+            return $this->error('Task cannot be found.', 404);
+        } catch (AuthorizationException $exception) {
+            return $this->error('You are not authorized to update this task.', 403);
         }
+
     }
 
     public function replace(ReplaceTaskRequest $request, $task_id) 
@@ -76,11 +85,15 @@ class TaskController extends ApiController
         try {
             $task = Task::findOrFail($task_id);
 
+            $this->authorize('replace', $task);
+
             $task->update($request->mappedAttributes());
 
             return new TaskResource($task);
         } catch (ModelNotFoundException $exception) {
-            return $this->error('Ticket cannot be found.', 404);
+            return $this->error('Task cannot be found.', 404);
+        } catch (AuthorizationException $exception) {
+            return $this->error('You are not authorized to replace this task.', 403);
         }
     }
 
@@ -91,11 +104,16 @@ class TaskController extends ApiController
     {
         try {
             $task = Task::findOrFail($task_id);
+
+            $this->authorize('delete', $task);
             $task->delete();
 
             return $this->success('task successfully deleted');
         } catch (ModelNotFoundException $exception) {
             return $this->error('task can not be found.', 404);
+        } catch (AuthorizationException $exception) {
+            return $this->error('You are not authorized to delete this task.', 403);
         }
+
     }
 }
